packages/bh-sentinel-ml/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bh-sentinel-ml"
version = "0.2.0"
description = "Transformer-based clinical safety signal detection for behavioral health systems"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "BH Healthcare", email = "oss@bh-healthcare.github.io" }
]
keywords = [
    "nlp",
    "clinical",
    "safety",
    "behavioral-health",
    "healthcare",
    "transformer",
    "onnx",
    "zero-shot",
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
    "Typing :: Typed",
]
dependencies = [
    "bh-sentinel-core>=0.1.1,<1",   # Pipeline + Flag model + taxonomy + preprocessor reuse
    "onnxruntime>=1.16,<2",         # In-process transformer inference (CPU provider)
    "tokenizers>=0.15,<1",          # HF tokenizers for NLI premise/hypothesis encoding
    "huggingface-hub>=0.20,<1",     # Pinned-revision model download from HF Hub
    "platformdirs>=4,<5",           # User-level cache dir resolution for downloaded models
    "pyyaml>=6,<7",                 # ml_config.yaml + zero_shot_hypotheses.yaml loaders
]

[project.scripts]
bh-sentinel-ml = "bh_sentinel.ml.cli.__main__:main"

[project.optional-dependencies]
eval = [
    "numpy>=1.24,<2",               # Calibration math, softmax, ECE computation
    "scikit-learn>=1.3,<2",         # Evaluation metrics (precision, recall, F1)
]
dev = [
    "pytest>=8,<9",                 # Test runner; compat-release; >=8 required by pytest-asyncio 1.x
    "ruff>=0.1.0,<1",               # Lint + format; 0.x cadence
    "mypy>=1.0,<2",                 # Strict type checking (see [tool.mypy] below)
    "onnx>=1.15,<2",                # Build the tiny NLI fixture in tests/conftest.py
]

[project.urls]
Homepage = "https://github.com/bh-healthcare/bh-sentinel"
Documentation = "https://github.com/bh-healthcare/bh-sentinel#readme"
Repository = "https://github.com/bh-healthcare/bh-sentinel"
Issues = "https://github.com/bh-healthcare/bh-sentinel/issues"
Changelog = "https://github.com/bh-healthcare/bh-sentinel/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/bh_sentinel"]
only-include = ["src/bh_sentinel/ml"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
]

[tool.hatch.build]
artifacts = [
    "py.typed",
    "**/_default_config/*.yaml",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "UP",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "real_model: diagnostic tests that need a real (or tiny-stub) ONNX model; skipped in default CI",
]
addopts = "-m 'not real_model'"
